Drop entity from the schema
We've decided that the entity block served no good purpose, and already
couldn't be edited because we enforced that it was redundant with the title.

This also simplifies code to manipulate judgments.

We also fix the schema to prevent extra top-level properties.  A friendly
anonymous user found this flaw in the schema:
https://en.wikipedia.beta.wmflabs.org/w/index.php?title=JADE:Revision/376901&diff=prev&oldid=381117

Bug: T179301
Change-Id: Iea9b79e88056fc701e7150865824153948d0573d