Translatable entity type
We realized that the entity type being judged appears in the title, and
therefore should be translated.  This patch adds support.

Unfortunately, the entity type names are set in configuration, which doesn't
allow us to take advantage of existing translation infrastructure.  This was
done to prevent l10n-bot from causing dramatic namespace renaming.  See also
the -".namespaces.php" files, which are a known wart solving the same problem.

Note that TitleHelper isn't used until the API is merged, but fits into
review here.

Change-Id: I67e3c75985269a5d740830443851198863786d71