Drop page judgments for this release
We haven't finished designing the "drafttopic" schema, plus there are difficult
questions still open about how to represent page judgments.  For example, one
suggestion is that they are stored using MCR in a non-main slot on the page
itself.  This is an interesting idea, but we're postponing in order to not
distract from core functionality.

Bug: T202596
Change-Id: I83c46ed535adbf5702c535d0f0458dc30b081339