Fix anonymous user id in DeleteEndorsement comment
The $userdata here is either an id for a logged in user
or a string containing an ip address for anonymous users.
If the user is not logged in when deleting an endorsement,
we set the $userid as zero and show that in the revision comment
instead of the ip address.

Bug: T245311
Change-Id: I75749617c6e84dfc12c1c7a29dbf4d512fcaf20b