Fix double endorsement on facet
User A was able to endorse a proposal in a facet, then was also able to create and endorse a new proposal.
This caused User A to endorse 2 proposals in the same facet.
This patch fixes this issue. Now User A is not able to create and endorse a new proposal if they've already endorsed a proposal in the same facet.

Bug: T247974
Change-Id: I1084409b698e158407ae26e549ea519c1d0571be