Localized Jade history page comment facet and labeldata parts
Forexample:
If an edit comment on the history page was;
    (Preference set: editquality is {"damaging":true,"goodfaith":true}: Should be numero uno!)
It is now;
    (Preference set: Edit quality is Damaging / Good-faith: Should be numero uno!)
The "editquality is" i.e facet part of the comment is now localized.
The "{"damaging":true,"goodfaith":true}" i.e labeldata part of the comment is now localized and this has been done for all the 3 labeldata types.

Bug: T247457
Change-Id: Iec9a78ed3ed4b7d77b2b114cbfb8967473b08e03